improve register banking, improve some arm instructions

+++ egg-vance/src/arm/arm.h
@@ -69,8 +69,8 @@ private:
     void unconditionalBranch(u16 instr);
     void longBranchLink(u16 instr);
 
-    u32 rotatedImmediate(u16 value, bool& carry);
-    u32 shiftedRegister(u16 value, bool& carry);
+    u32 rotatedImmediate(int value, bool& carry);
+    u32 shiftedRegister(int value, bool& carry);
 
     void branchExchange(u32 instr);
     void branchLink(u32 instr);
